refactor: Update import paths for event bus and message types; enhance VibeCoderChat and VibeCoderControls functionality
